split into ExtendedMediaEntity
