+++ twitter4j-appengine/src/main/java/twitter4j/LazyDirectMessage.java
@@ -129,7 +129,7 @@ final class LazyDirectMessage implements twitter4j.DirectMessage {
     }
 
     @Override
-    public MediaEntity[] getExtendedMediaEntities() {
+    public ExtendedMediaEntity[] getExtendedMediaEntities() {
         return getTarget().getExtendedMediaEntities();
     }
