+++ twitter4j-appengine/src/main/java/twitter4j/LazyStatus.java
@@ -318,7 +318,7 @@ final class LazyStatus implements twitter4j.Status {
     }
     
     @Override
-    public MediaEntity[] getExtendedMediaEntities() {
+    public ExtendedMediaEntity[] getExtendedMediaEntities() {
         return getTarget().getExtendedMediaEntities();
     }
