+++ twitter4j-core/src/main/java/twitter4j/MediaEntity.java
@@ -71,21 +71,4 @@ public interface MediaEntity extends URLEntity {
      * @return the media type ("photo", "video", "animated_gif").
      */
     String getType();
-
-    int getVideoAspectRatioWidth();
-
-    int getVideoAspectRatioHeight();
-
-    long getVideoDurationMillis();
-
-    interface Variant extends java.io.Serializable {
-
-        int getBitrate();
-
-        String getContentType();
-
-        String getUrl();
-    }
-
-    Variant[] getVideoVariants();
 }
